requirements.txt
PyGithub==2.6.1
pytest==8.3.5